+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 - Adds support for @query type (user defined functions)
 - Adds support for recursive references
 - Removes `get`, `post`, `put`, `patch` and `delete` methods from `Client` class
+- Adds `abort` function
+- Adds `normalizer` argument to `casefold` function
+- Adds `new_id` function
+- Deprecated `next_id` in favor of `new_id`
+- Adds `identity` and `has_identity` functions
+- Adds `singleton` and `events` functions
 
 ## 1.1.2 (July 06, 2017)
 - Fix typescript declaration file to include default types